Refactor: StyledComponentInSetting import 추가, 모달창 props 설정
