--- src/components/FinancialLedger/SettingListContentWrapper.js
@@ -1,25 +1,0 @@
-import styled from 'styled-components';
-
-//리스트 컴포넌트 스타일링
-export const SettingList = styled.div`
-    margin: 3em;
-`;
-
-//설정 항목 제목 컴포넌트 스타일링
-export const SettingListTitleWrapper = styled.div`
-        font-size: 20px;
-        font-weight: bold;
-        display: flex;
-        justify-content: space-between;
-        width: 49%;
-`;
-
-//설정 항목 내용 컴포넌트 스타일링
-export const SettingListContentWrapper = styled.div`
-        font-size: 17px;
-        font-weight: normal;
-        margin: 1em;
-        display: flex;
-        justify-content: space-between;
-        width: 49%;
-    `;
